back/requirements.txt
annotated-types==0.7.0
anyio==4.3.0
asgiref==3.7.2
asttokens==2.4.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
dj-rest-auth==5.0.2
Django==4.2.8
django-allauth==0.59.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.14.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.1.1
ipykernel==6.28.0
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
Markdown==3.6
matplotlib==3.8.2
matplotlib-inline==0.1.6
mysqlclient==2.2.4
nest-asyncio==1.5.8
numpy==1.26.3
oauthlib==3.2.2
openai==1.30.1
pillow==10.3.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
sniffio==1.3.1
sqlparse==0.4.4
tqdm==4.66.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.13
zipp==3.17.0
